requirements.txt
flask==2.3.3
redis==5.0.1
pymysql==1.1.1
requests==2.32.4
pycryptodome==3.19.0
python-dotenv==1.0.0